fix: updated tests
