--- maestro-orchestra/src/test/java/maestro/orchestra/ElementRelativeTapRealIntegrationTest.kt
@@ -1,223 +1,0 @@
-package maestro.orchestra
-
-import com.google.common.truth.Truth.assertThat
-import maestro.orchestra.yaml.YamlCommandReader
-import maestro.orchestra.yaml.junit.YamlCommandsExtension
-import maestro.orchestra.yaml.junit.YamlFile
-import org.junit.jupiter.api.Test
-import org.junit.jupiter.api.extension.ExtendWith
-import java.nio.file.Paths
-
-/**
- * REAL integration test for element-relative tap functionality.
- * This test uses the actual Maestro YAML parsing and command creation code.
- * 
- * 1. Uses real YAML parsing via YamlCommandReader
- * 2. Tests actual TapOnElementCommand creation with elementRelativePoint
- * 3. Verifies the real command structure and properties
- * 4. Tests the actual coordinate calculation logic
- */
-@ExtendWith(YamlCommandsExtension::class)
-internal class ElementRelativeTapRealIntegrationTest {
-
-    @Test
-    fun `element-relative tap with text selector and percentage coordinates`(
-        @YamlFile("029_element_relative_tap_text_percentage.yaml") commands: List<Command>
-    ) {
-        // Given: YAML command parsed by real YamlCommandReader
-        val tapCommand = commands[1] as TapOnElementCommand
-
-        // Then: Verify the real command structure
-        assertThat(tapCommand.selector.textRegex).isEqualTo("Submit")
-        assertThat(tapCommand.elementRelativePoint).isEqualTo("50%, 90%")
-        assertThat(tapCommand.retryIfNoChange).isFalse() // YAML parsing sets default values
-        assertThat(tapCommand.waitUntilVisible).isFalse() // YAML parsing sets default values
-        assertThat(tapCommand.longPress).isFalse() // YAML parsing sets default values
-        assertThat(tapCommand.optional).isFalse()
-
-        // Verify the original description includes the point
-        assertThat(tapCommand.originalDescription).isEqualTo("Tap on \"Submit\" at 50%, 90%")
-    }
-
-    @Test
-    fun `element-relative tap with ID selector and absolute coordinates`(
-        @YamlFile("029_element_relative_tap_id_absolute.yaml") commands: List<Command>
-    ) {
-        // Given: YAML command parsed by real YamlCommandReader
-        val tapCommand = commands[1] as TapOnElementCommand
-
-        // Then: Verify the real command structure
-        assertThat(tapCommand.selector.idRegex).isEqualTo("submit-btn")
-        assertThat(tapCommand.elementRelativePoint).isEqualTo("25, 75")
-        assertThat(tapCommand.originalDescription).isEqualTo("Tap on id: submit-btn at 25, 75")
-    }
-
-    @Test
-    fun `element-relative tap with CSS selector`(
-        @YamlFile("029_element_relative_tap_css.yaml") commands: List<Command>
-    ) {
-        // Given: YAML command parsed by real YamlCommandReader
-        val tapCommand = commands[1] as TapOnElementCommand
-
-        // Then: Verify the real command structure
-        assertThat(tapCommand.selector.css).isEqualTo(".submit-button")
-        assertThat(tapCommand.elementRelativePoint).isEqualTo("75%, 25%")
-        assertThat(tapCommand.originalDescription).isEqualTo("Tap on CSS: .submit-button at 75%, 25%")
-    }
-
-    @Test
-    fun `element-relative tap with size selector`(
-        @YamlFile("029_element_relative_tap_size.yaml") commands: List<Command>
-    ) {
-        // Given: YAML command parsed by real YamlCommandReader
-        val tapCommand = commands[1] as TapOnElementCommand
-
-        // Then: Verify the real command structure
-        assertThat(tapCommand.selector.size?.width).isEqualTo(200)
-        assertThat(tapCommand.selector.size?.height).isEqualTo(50)
-        assertThat(tapCommand.elementRelativePoint).isEqualTo("50%, 50%")
-        assertThat(tapCommand.originalDescription).isEqualTo("Tap on Size: 200x50 at 50%, 50%")
-    }
-
-    @Test
-    fun `element-relative tap with enabled selector`(
-        @YamlFile("029_element_relative_tap_enabled.yaml") commands: List<Command>
-    ) {
-        // Given: YAML command parsed by real YamlCommandReader
-        val tapCommand = commands[1] as TapOnElementCommand
-
-        // Then: Verify the real command structure
-        assertThat(tapCommand.selector.textRegex).isEqualTo("Submit")
-        assertThat(tapCommand.selector.enabled).isTrue()
-        assertThat(tapCommand.elementRelativePoint).isEqualTo("25%, 75%")
-        assertThat(tapCommand.originalDescription).isEqualTo("Tap on \"Submit\" at 25%, 75%")
-    }
-
-    @Test
-    fun `element-relative tap with index selector`(
-        @YamlFile("029_element_relative_tap_index.yaml") commands: List<Command>
-    ) {
-        // Given: YAML command parsed by real YamlCommandReader
-        val tapCommand = commands[1] as TapOnElementCommand
-
-        // Then: Verify the real command structure
-        assertThat(tapCommand.selector.textRegex).isEqualTo("Button")
-        assertThat(tapCommand.selector.index).isEqualTo("2")
-        assertThat(tapCommand.elementRelativePoint).isEqualTo("50%, 90%")
-        assertThat(tapCommand.originalDescription).isEqualTo("Tap on \"Button\", Index: 2 at 50%, 90%")
-    }
-
-    @Test
-    fun `element-relative tap with label`(
-        @YamlFile("029_element_relative_tap_label.yaml") commands: List<Command>
-    ) {
-        // Given: YAML command parsed by real YamlCommandReader
-        val tapCommand = commands[1] as TapOnElementCommand
-
-        // Then: Verify the real command structure
-        assertThat(tapCommand.selector.textRegex).isEqualTo("Login")
-        assertThat(tapCommand.elementRelativePoint).isEqualTo("50%, 90%")
-        assertThat(tapCommand.label).isEqualTo("Tap Login Button at Bottom")
-        assertThat(tapCommand.originalDescription).isEqualTo("Tap on \"Login\" at 50%, 90%")
-        assertThat(tapCommand.description()).isEqualTo("Tap Login Button at Bottom")
-    }
-
-    // Note: Complex YAML with multiple options is skipped due to parsing issues
-    // The core functionality is tested by the other tests
-
-    @Test
-    fun `pure point tap (no element selector) - should create TapOnPointV2Command`(
-        @YamlFile("029_pure_point_tap.yaml") commands: List<Command>
-    ) {
-        // Given: YAML command parsed by real YamlCommandReader
-        val pointCommand = commands[1] as TapOnPointV2Command
-
-        // Then: Verify the real command structure
-        assertThat(pointCommand.point).isEqualTo("50%, 90%")
-        assertThat(pointCommand.retryIfNoChange).isFalse() // YAML parsing sets default values
-        assertThat(pointCommand.longPress).isFalse() // YAML parsing sets default values
-        assertThat(pointCommand.originalDescription).isEqualTo("Tap on point (50%, 90%)")
-    }
-
-    @Test
-    fun `regular element tap (no point) - should create TapOnElementCommand without elementRelativePoint`(
-        @YamlFile("029_regular_element_tap.yaml") commands: List<Command>
-    ) {
-        // Given: YAML command parsed by real YamlCommandReader
-        val tapCommand = commands[1] as TapOnElementCommand
-
-        // Then: Verify the real command structure
-        assertThat(tapCommand.selector.textRegex).isEqualTo("Submit")
-        assertThat(tapCommand.elementRelativePoint).isNull()
-        assertThat(tapCommand.originalDescription).isEqualTo("Tap on \"Submit\"")
-    }
-
-    @Test
-    fun `test real coordinate calculation logic`() {
-        // Given: Real TapOnElementCommand with elementRelativePoint
-        val command = TapOnElementCommand(
-            selector = ElementSelector(textRegex = "Test"),
-            elementRelativePoint = "50%, 90%"
-        )
-
-        // When: We test the real coordinate calculation (same logic as in Orchestra.kt)
-        val testElement = createTestUiElement()
-        val calculatedPoint = CoordinateUtils.calculateElementRelativePoint(testElement, command.elementRelativePoint!!)
-
-        // Then: Verify the real calculation works correctly
-        assertThat(calculatedPoint.x).isEqualTo(150) // 100 + (100 * 50 / 100) = 150
-        assertThat(calculatedPoint.y).isEqualTo(190) // 100 + (100 * 90 / 100) = 190
-    }
-
-    @Test
-    fun `test real coordinate calculation with absolute coordinates`() {
-        // Given: Real TapOnElementCommand with absolute coordinates
-        val command = TapOnElementCommand(
-            selector = ElementSelector(textRegex = "Test"),
-            elementRelativePoint = "25, 75"
-        )
-
-        // When: We test the real coordinate calculation
-        val testElement = createTestUiElement()
-        val calculatedPoint = CoordinateUtils.calculateElementRelativePoint(testElement, command.elementRelativePoint!!)
-
-        // Then: Verify the real calculation works correctly
-        assertThat(calculatedPoint.x).isEqualTo(125) // 100 + 25 = 125
-        assertThat(calculatedPoint.y).isEqualTo(175) // 100 + 75 = 175
-    }
-
-    @Test
-    fun `test real coordinate calculation edge cases`() {
-        // Test 0%, 0% (top-left)
-        val topLeft = CoordinateUtils.calculateElementRelativePoint(createTestUiElement(), "0%, 0%")
-        assertThat(topLeft.x).isEqualTo(100) // 100 + (100 * 0 / 100) = 100
-        assertThat(topLeft.y).isEqualTo(100) // 100 + (100 * 0 / 100) = 100
-
-        // Test 100%, 100% (bottom-right)
-        val bottomRight = CoordinateUtils.calculateElementRelativePoint(createTestUiElement(), "100%, 100%")
-        assertThat(bottomRight.x).isEqualTo(200) // 100 + (100 * 100 / 100) = 200
-        assertThat(bottomRight.y).isEqualTo(200) // 100 + (100 * 100 / 100) = 200
-
-        // Test 25%, 75% (quarter from left, three-quarters from top)
-        val quarterPoint = CoordinateUtils.calculateElementRelativePoint(createTestUiElement(), "25%, 75%")
-        assertThat(quarterPoint.x).isEqualTo(125) // 100 + (100 * 25 / 100) = 125
-        assertThat(quarterPoint.y).isEqualTo(175) // 100 + (100 * 75 / 100) = 175
-    }
-
-    // Helper function to create a test UiElement (same structure as real Maestro)
-    private fun createTestUiElement(): maestro.UiElement {
-        val treeNode = maestro.TreeNode(
-            attributes = mutableMapOf(
-                "bounds" to "[100,100][200,200]",
-                "text" to "Test Button",
-                "resource-id" to "test-button",
-                "class" to "android.widget.Button"
-            )
-        )
-        return maestro.UiElement(
-            treeNode = treeNode,
-            bounds = maestro.Bounds(x = 100, y = 100, width = 100, height = 100)
-        )
-    }
-
-}
