+++ maestro-orchestra/src/test/java/maestro/orchestra/yaml/YamlCommandReaderTest.kt
@@ -586,6 +586,136 @@ internal class YamlCommandReaderTest {
         )
     }
 
+    // Element-relative tap tests
+    @Test
+    fun `element-relative tap with text selector and percentage coordinates`(
+        @YamlFile("029_element_relative_tap_text_percentage.yaml") commands: List<Command>
+    ) {
+        // Given: YAML command parsed by real YamlCommandReader
+        val tapCommand = commands[1] as TapOnElementCommand
+
+        // Then: Verify the real command structure
+        assertThat(tapCommand.selector.textRegex).isEqualTo("Submit")
+        assertThat(tapCommand.elementRelativePoint).isEqualTo("50%, 90%")
+        assertThat(tapCommand.retryIfNoChange).isFalse() // YAML parsing sets default values
+        assertThat(tapCommand.waitUntilVisible).isFalse() // YAML parsing sets default values
+        assertThat(tapCommand.longPress).isFalse() // YAML parsing sets default values
+        assertThat(tapCommand.optional).isFalse()
+
+        // Verify the original description includes the point
+        assertThat(tapCommand.originalDescription).isEqualTo("Tap on \"Submit\" at 50%, 90%")
+    }
+
+    @Test
+    fun `element-relative tap with ID selector and absolute coordinates`(
+        @YamlFile("029_element_relative_tap_id_absolute.yaml") commands: List<Command>
+    ) {
+        // Given: YAML command parsed by real YamlCommandReader
+        val tapCommand = commands[1] as TapOnElementCommand
+
+        // Then: Verify the real command structure
+        assertThat(tapCommand.selector.idRegex).isEqualTo("submit-btn")
+        assertThat(tapCommand.elementRelativePoint).isEqualTo("25, 75")
+        assertThat(tapCommand.originalDescription).isEqualTo("Tap on id: submit-btn at 25, 75")
+    }
+
+    @Test
+    fun `element-relative tap with CSS selector`(
+        @YamlFile("029_element_relative_tap_css.yaml") commands: List<Command>
+    ) {
+        // Given: YAML command parsed by real YamlCommandReader
+        val tapCommand = commands[1] as TapOnElementCommand
+
+        // Then: Verify the real command structure
+        assertThat(tapCommand.selector.css).isEqualTo(".submit-button")
+        assertThat(tapCommand.elementRelativePoint).isEqualTo("75%, 25%")
+        assertThat(tapCommand.originalDescription).isEqualTo("Tap on CSS: .submit-button at 75%, 25%")
+    }
+
+    @Test
+    fun `element-relative tap with size selector`(
+        @YamlFile("029_element_relative_tap_size.yaml") commands: List<Command>
+    ) {
+        // Given: YAML command parsed by real YamlCommandReader
+        val tapCommand = commands[1] as TapOnElementCommand
+
+        // Then: Verify the real command structure
+        assertThat(tapCommand.selector.size?.width).isEqualTo(200)
+        assertThat(tapCommand.selector.size?.height).isEqualTo(50)
+        assertThat(tapCommand.elementRelativePoint).isEqualTo("50%, 50%")
+        assertThat(tapCommand.originalDescription).isEqualTo("Tap on Size: 200x50 at 50%, 50%")
+    }
+
+    @Test
+    fun `element-relative tap with enabled selector`(
+        @YamlFile("029_element_relative_tap_enabled.yaml") commands: List<Command>
+    ) {
+        // Given: YAML command parsed by real YamlCommandReader
+        val tapCommand = commands[1] as TapOnElementCommand
+
+        // Then: Verify the real command structure
+        assertThat(tapCommand.selector.textRegex).isEqualTo("Submit")
+        assertThat(tapCommand.selector.enabled).isTrue()
+        assertThat(tapCommand.elementRelativePoint).isEqualTo("25%, 75%")
+        assertThat(tapCommand.originalDescription).isEqualTo("Tap on \"Submit\" at 25%, 75%")
+    }
+
+    @Test
+    fun `element-relative tap with index selector`(
+        @YamlFile("029_element_relative_tap_index.yaml") commands: List<Command>
+    ) {
+        // Given: YAML command parsed by real YamlCommandReader
+        val tapCommand = commands[1] as TapOnElementCommand
+
+        // Then: Verify the real command structure
+        assertThat(tapCommand.selector.textRegex).isEqualTo("Button")
+        assertThat(tapCommand.selector.index).isEqualTo("2")
+        assertThat(tapCommand.elementRelativePoint).isEqualTo("50%, 90%")
+        assertThat(tapCommand.originalDescription).isEqualTo("Tap on \"Button\", Index: 2 at 50%, 90%")
+    }
+
+    @Test
+    fun `element-relative tap with label`(
+        @YamlFile("029_element_relative_tap_label.yaml") commands: List<Command>
+    ) {
+        // Given: YAML command parsed by real YamlCommandReader
+        val tapCommand = commands[1] as TapOnElementCommand
+
+        // Then: Verify the real command structure
+        assertThat(tapCommand.selector.textRegex).isEqualTo("Login")
+        assertThat(tapCommand.elementRelativePoint).isEqualTo("50%, 90%")
+        assertThat(tapCommand.label).isEqualTo("Tap Login Button at Bottom")
+        assertThat(tapCommand.originalDescription).isEqualTo("Tap on \"Login\" at 50%, 90%")
+        assertThat(tapCommand.description()).isEqualTo("Tap Login Button at Bottom")
+    }
+
+    @Test
+    fun `pure point tap (no element selector) - should create TapOnPointV2Command`(
+        @YamlFile("029_pure_point_tap.yaml") commands: List<Command>
+    ) {
+        // Given: YAML command parsed by real YamlCommandReader
+        val pointCommand = commands[1] as TapOnPointV2Command
+
+        // Then: Verify the real command structure
+        assertThat(pointCommand.point).isEqualTo("50%, 90%")
+        assertThat(pointCommand.retryIfNoChange).isFalse() // YAML parsing sets default values
+        assertThat(pointCommand.longPress).isFalse() // YAML parsing sets default values
+        assertThat(pointCommand.originalDescription).isEqualTo("Tap on point (50%, 90%)")
+    }
+
+    @Test
+    fun `regular element tap (no point) - should create TapOnElementCommand without elementRelativePoint`(
+        @YamlFile("029_regular_element_tap.yaml") commands: List<Command>
+    ) {
+        // Given: YAML command parsed by real YamlCommandReader
+        val tapCommand = commands[1] as TapOnElementCommand
+
+        // Then: Verify the real command structure
+        assertThat(tapCommand.selector.textRegex).isEqualTo("Submit")
+        assertThat(tapCommand.elementRelativePoint).isNull()
+        assertThat(tapCommand.originalDescription).isEqualTo("Tap on \"Submit\"")
+    }
+
     private fun commands(vararg commands: Command): List<MaestroCommand> =
         commands.map(::MaestroCommand).toList()
 }
